project structure refactoring. util
