+++ mdb/lib/src/data/api/api.dart
@@ -3,7 +3,7 @@ import 'package:dio/dio.dart';
 import 'package:mdb/src/data/model/remote/discover_response.dart';
 import 'package:mdb/src/data/model/remote/genres_response.dart';
 import 'package:mdb/src/data/model/remote/popular_response.dart';
-import 'package:mdb/src/utils/constants.dart';
+import 'package:mdb/src/util/constants.dart';
 
 class Api {
   Dio dio = Dio();
