+++ mdb/lib/src/data/model/remote/popular_response.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:mdb/src/data/model/local/movie.dart';
-import 'package:mdb/src/utils/constants.dart';
+import 'package:mdb/src/util/constants.dart';
 
 //part 'popular_response.g.dart';
