+++ mdb/lib/src/ui/home_screen/dicover_screen.dart
@@ -6,7 +6,7 @@ import 'package:mdb/src/data/model/local/movie.dart';
 import 'package:mdb/src/data/model/remote/discover_response.dart';
 import 'package:mdb/src/data/model/remote/genres_response.dart';
 import 'package:mdb/src/ui/movie_page_viver_item.dart';
-import 'package:mdb/src/utils/wigdet/page_transformer.dart';
+import 'package:mdb/src/util/wigdet/page_transformer.dart';
 
 class DiscoverScreen extends StatefulWidget {
   @override
